20/11/2023 00:40

+++ src/main/java/com/piltover/entity/Booking.java
@@ -51,7 +51,7 @@ public class Booking implements Serializable {
     private Discount discount;
 
     @Column(name = "Total_price")
-    private Integer totalPrice;
+    private Double totalPrice;
 
     @Column(name = "Total_passengers")
     private Integer totalPassengers;
